+++ NEWS.md
@@ -4,10 +4,20 @@
 * 
 
 ### Changed
-* 
+
+* w.obs -> w_obs
+* inv.Sigma.obs -> inv_sigma_obs
+* obs.use -> scaled_obs
+* tune.fn -> compute_m_sigma
+* GP.weights.test -> compute_weight_gp
+* data.generate -> generate_synthetic_data 
+
 
 ### Added
-* 
+
+* estimate_cerf_gp function
+* compute_inverse function
+* compute_w_corr function
 
 ### Removed
 *